pyproject.toml
[project]
name = "camelai-hw3"
version = "0.1.0"
description = ""
authors = [
    {name = "Jyrko",email = "uriydemoshenko@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "camel-ai[all] (>=0.2.62,<0.3.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "humanlayer (>=0.7.9,<0.8.0)",
    "nest-asyncio (>=1.6.0,<2.0.0)"
]

[tool.poetry]
packages = [
    { include = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
